Extrair view_item_sumario_agenda.xml
