+++ mercos/src/main/java/com/github/felipeslongo/honeycomb/mercos/agenda/ItemSumarioAgendaViewState.kt
@@ -1,0 +1,24 @@
+package com.github.felipeslongo.honeycomb.mercos.agenda
+
+data class ItemSumarioAgendaViewState(
+    val id: Int,
+    val nomeDoCliente: String = "",
+    val nomeDoRoteiro: String = "",
+    val horario: String = "",
+    val nomeDoResponsavel: String = "",
+    val telefone: String = "",
+    val endereco: String = ""
+) {
+
+    companion object {
+        fun createLoremIpsum(id: Int) = ItemSumarioAgendaViewState(
+            id,
+            "Lima Luminárias e Acessórios",
+            "Rota Zona Sul",
+            "15:32",
+            "João Martins",
+            "9 8874-1218",
+            "Rua Benjamin Constant, 9876"
+        )
+    }
+}
