+++ mercos/src/main/res/layout/view_item_sumario_agenda.xml
@@ -1,0 +1,88 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+        <variable
+                name="viewModel"
+                type="com.github.felipeslongo.honeycomb.mercos.agenda.ItemSumarioAgendaViewModel" />
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/constraintLayout_sumario"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content">
+
+        <ImageView
+                android:id="@+id/imageView_tipo"
+                android:layout_width="40dp"
+                android:layout_height="40dp"
+                android:layout_gravity="center"
+                android:background="@drawable/shape_circle"
+                app:backgroundTint="@color/mercos_cinza_100"
+                android:padding="8dp"
+                app:srcCompat="@drawable/ic_baseline_location_on_24"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent" />
+
+        <TextView
+                android:id="@+id/textView_cliente_nome"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="8dp"
+                android:layout_marginLeft="8dp"
+                android:text="Cliente"
+                android:textSize="16sp"
+                android:textStyle="bold"
+                app:layout_constraintStart_toEndOf="@+id/imageView_tipo"
+                app:layout_constraintTop_toTopOf="parent"
+                tools:text="Lima Luminárias e Acessórios" />
+
+        <TextView
+                android:id="@+id/textView_roteiro"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="4dp"
+                android:layout_marginLeft="4dp"
+                android:text="Roteiro"
+                app:layout_constraintBottom_toBottomOf="@+id/imageView_roteiro"
+                app:layout_constraintStart_toEndOf="@+id/imageView_roteiro"
+                app:layout_constraintTop_toTopOf="@+id/imageView_roteiro"
+                tools:text="Rota Zona Sul" />
+
+        <ImageView
+                android:id="@+id/imageView_roteiro"
+                android:layout_width="17dp"
+                android:layout_height="17dp"
+                android:layout_marginTop="4dp"
+                android:tint="@color/mercos_cinza_500"
+                app:layout_constraintStart_toStartOf="@+id/textView_cliente_nome"
+                app:layout_constraintTop_toBottomOf="@+id/textView_cliente_nome"
+                app:srcCompat="@drawable/ic_baseline_directions_24" />
+
+        <TextView
+                android:id="@+id/textView_horario"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="4dp"
+                android:text="às 00:00"
+                android:textSize="12sp"
+                app:layout_constraintStart_toStartOf="@+id/textView_cliente_nome"
+                app:layout_constraintTop_toBottomOf="@+id/imageView_roteiro"
+                tools:text="às 15:32" />
+
+        <TextView
+                android:id="@+id/textView_responsavel"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="8dp"
+                android:layout_marginLeft="8dp"
+                android:text="Responsável"
+                android:textSize="12sp"
+                app:layout_constraintStart_toEndOf="@+id/textView_horario"
+                app:layout_constraintTop_toTopOf="@+id/textView_horario"
+                tools:text="João Martins" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
